workers/requirements.txt
supabase>=2.8.0
httpx>=0.27.0
pandas>=2.2.0
pytrends>=4.9.2
python-dotenv>=1.0.1
